Drivers/HAL: Add a drivers package
`drivers` is a new hardware abstraction layer for reef-pi, which
provides a pluggable interface and consolidation for the variety of
modules in the reef-pi code base. A driver is designed to provide a
named entry point, and can have a variety of interfaces representing
those capabilities. For example, Input, Output, PWM, Temperature, pH,
etc.