pH calibration check frequency
You probably remember this was changed to 1000ms, that was still too often for the Atlas circuit and errors appeared after 4 minutes of running. The absolute minimum appears to be 1070ms but I think 1500 guarantees there will be no timing issues. I have fully testing 1500ms and it works perfect.