First cut of a German translation
This is the main UI, including additions and fixes in .jsx affecting other languages.
There are still lots of validation messages literally in the *schema*
files.

One suggested feature change: on temperature chart I changed the legend
box to show usage percentages for heater/chjiller instead of seconds. (This might be
nicer for the data series in the chart itself too.) The 100% reference
is a constant of 3600, which may be wrong for a different time scale.
Don't where to get this value dynamically.

This is my first bigger commit to this projct, I therefore don't know whether go.mod, go.sum and package-lock.json should be included. I did not actively modify them.