Move all dependencies to the dependencies section
There is no reason to differentiate between dev dependencies and regular
dependencies in a React project. All dependencies are used to build
artifacts that should be served to a browser.

I guess one could argue that all are dev dependencies. But this is where
Create React App places them. So I reasoned that it was better to go
with the established convention.