Remove electron from dependencies
Electron is a quite large dependency, that I rather not have to download
and build during `yarn install`. And given that we aren't using it
directly I see no reason to keep it around.

This removes electron as a direct dependency, though it is still present
as a transitive dependency. So further work is needed to remove it
completely.