Macro validation
Changed to use generic validation messages.

Added translation of subsystems, which hopefully doesn't break functionality.

The new string group 'function:*' will be used again (and extended) by the
coming PR for timers.
Long term I'd want to consolidate the substem names as used on tab titles,
checkboxes of Settings-Capabilities, and in macros and timers to the two
string groups:
- capabilities:* for the cathegory name (tab and caps.) and
- function:* for the idividual controlled item.
This will result in a more consistent terminology throughtout the UI.