Support direct links to pages
This replaces the redux / custom management of the pages with a React Router
based version. The main impact of this is now you can directly go to urls
ala `/temperature` and it will render the appropriate page. The nav links
also point to this, which means you can also "open in new tab" to have
multiple reef pi tabs working at once.

I have tested this by going through the various pages and adding/deleting
things, playing with the forms, and navigating to various pages. It all
works for me. However I'm not sure if there's existing precedent for auto
testing this sort of change in this app.

I would recommend poking around at it a bit yourself as well to triple
check I didn't miss some behavior change, but it worked in my testing.

Implements #1780

Note while implementing this I found #1842, which I didn't try and fix in
this commit. Mostly because I'm not sure how to implement it in React, and I
gave up trying for the moment.