Support direct links to configuration
Similar to the other links, but allows for deep linking to specific config
pages.

Also need to set the web pack location to an absolute path to make sure
it doesn't render as a nested url (eg /configuration.../assets).

Relates to #1779 in that adding this will make it easy to direct link to the
error screen.