Link the Errors in footer to the Errors page
This is built on top of #1843. This doesn't combine the Log and Error screens,
but instead deep links to the error page.

I believe combining the Log and Errors screens will still be worthwhile, but
this is an easy improvement with the deep linking in place.

Tested just by pulling up the UI and seeing the link works.

Relates #1779