Handle rendering the connectors screen before drivers are loaded
This loads the drivers explicitly in the parent, because multiple children need
them. It might be more consistent with other components to also have each child
load the drivers, but this felt more efficient.

I think an argument could also be made to have the parent inject the drivers
into the children, but I was leery that there may be some unexpected side
effect to that. So I'm just going with this version.