Don't require a heater or cooler temp when they're disabled
Fixes #1836

TBH I'm not sure exactly what the behavior of hysteresis is, so I'm not
sure if there's cases where this validation is erroneous. However, this
fixes the bug where right now if you try and create a new temperature
with just a heater, it will fail validation without telling you why. Currently
it's requiring you to setup a cooler regardless, as outlined in #1836
e Please enter the commit message for your changes. Lines starting