Fix the Dockerfile to allow building
I'm having issues building reef-pi on an ARM mac. To workaround that I'm trying
to build under Docker. It appears this Dockerfile is old, and possibly not been
getting used. For example the node version it was using is very dated.

This updates the Dockerfile so at a minimum I am able to build the docker image
and within it build deb:

```
\# Build docker image and run it
$ docker build --platform=linux/amd64 -t debugreef
$ docker run -it debugreef bash

\# Within that image build reef-pi
$ make bin
$ make deb
...
Created package {:path=>"reef-pi-5.2-72-g7e5eeb6e.deb"}
```