Continue on partial failure in temp control
<WIP NOTE>
This is a WIP. I haven't tested it locally yet with my actual reef-pi, nor do
I know how to write an automated test for it. I don't see any controller tests
when I did a quick check.

Sending as a WIP to give a heads-up on the fix.
</WIP NOTE>

This updates the control code to flip as many heater/coolers as possible into
the correct state when performing an action, regardless of what happens to some
of them. For instance, if we're trying to heat the tank, we should try and turn
on the heater regardless of what happens to the cooler. Similarly that same
change applies to flipping on a cooler or turning everything off.

It's generally important this tries to flip everything to the proper state
versus just reordering things. For instance, if we're trying to heat the tank
we should flip the cooler off regardless of what happens to the heater and flip
the heater on regardless of the cooler.

Fixes #2013